+++ src/lib/openjp2/jp2.h
@@ -366,7 +366,7 @@ bool opj_jp2_read_header(  opj_stream_private_t *p_stream,
  */
 bool opj_jp2_read_tile_header ( opj_jp2_t * p_jp2,
                                 uint32_t * p_tile_index,
-                                uint32_t * p_data_size,
+                                uint64_t * p_data_size,
                                 uint32_t * p_tile_x0,
                                 uint32_t * p_tile_y0,
                                 uint32_t * p_tile_x1,
@@ -389,7 +389,7 @@ bool opj_jp2_read_tile_header ( opj_jp2_t * p_jp2,
 bool opj_jp2_write_tile (  opj_jp2_t *p_jp2,
                            uint32_t p_tile_index,
                            uint8_t * p_data,
-                           uint32_t p_data_size,
+                           uint64_t p_data_size,
                            opj_stream_private_t *p_stream,
                            opj_event_mgr_t * p_manager );
 
@@ -407,7 +407,7 @@ bool opj_jp2_write_tile (  opj_jp2_t *p_jp2,
 bool opj_jp2_decode_tile (  opj_jp2_t * p_jp2,
                             uint32_t p_tile_index,
                             uint8_t * p_data,
-                            uint32_t p_data_size,
+                            uint64_t p_data_size,
                             opj_stream_private_t *p_stream,
                             opj_event_mgr_t * p_manager );
